functions/battery_message_store/requirements.unexpanded.txt
google-cloud-firestore==2.1.0